requirements.txt
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.1
python-dateutil>=2.9.0
lxml>=5.0.0
python-dotenv>=1.0.0
